Updated sample and testing program.
Grew array from 70000 integers to 150000 integers.

Removed commented out code to print out array.

Changed type of counters from int to long unsigned int to show greater
counts.

Added inbuilt qsort function to compare timer and counter.

+++ sort.c
@@ -10,7 +10,7 @@
 #define TRUE !FALSE
 #define SORT_COMBSORT_SHRINKFACTOR 1.3
 
-extern int swapcount;
+extern unsigned long int swapcount;
 
 /* Swap the specified amount of bytes between two memory areas. */
 static void memswap(void *ptr1, void *ptr2, size_t len) {
